refactor: narrow fallback helper boundaries
