+++ src/fallbacks.ts
@@ -1,17 +1,5 @@
 import type { PullRequestDraft } from "./types";
-import { clip, normalizeWhitespace, stableHash } from "./utils";
-
-function slugSegment(value: string): string {
-  const ascii = Array.from(value.normalize("NFKD"))
-    .filter((character) => character.charCodeAt(0) <= 0x7f)
-    .join("");
-
-  return ascii
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/-+/g, "-")
-    .replace(/^-|-$/g, "");
-}
+import { clip, normalizeWhitespace, slugSegment, stableHash } from "./utils";
 
 function shortSummary(prompt: string, maxLength = 60): string {
   return clip(normalizeWhitespace(prompt), maxLength);
@@ -59,32 +47,5 @@ export function fallbackPullRequestDraft(
       "## Notes",
       "- Generated by the deterministic Bun Codex harness.",
     ].join("\n"),
-  };
-}
-
-export function coerceBranchName(value: string): string {
-  const trimmed = value.trim().split(/\r?\n/, 1)[0] ?? "";
-  const normalized = trimmed
-    .replace(/^refs\/heads\//, "")
-    .replace(/[^A-Za-z0-9/_-]+/g, "-")
-    .replace(/\/+/g, "/")
-    .replace(/-+/g, "-")
-    .replace(/(^[/.]+|[/.]+$)/g, "");
-
-  return normalized;
-}
-
-export function parseDraftResponse(output: string): PullRequestDraft | null {
-  const normalized = output.replace(/\r\n/g, "\n").trim();
-  const titleMatch = normalized.match(/^TITLE:\s*(.+)$/m);
-  const bodyMatch = normalized.match(/^BODY:\s*([\s\S]*)$/m);
-
-  if (!titleMatch || !bodyMatch) {
-    return null;
-  }
-
-  return {
-    title: titleMatch[1]?.trim() ?? "",
-    body: bodyMatch[1]?.trim() ?? "",
   };
 }
